Export Function Added and saprated home And History page and some ui fixes
