+++ ImListening/ClientApp/src/components/Home_style.css
@@ -1,62 +1,1 @@
-﻿:root {
-    --bgColor : #F3F3F3;
-    --textColor : #424242;
-    --whiteColor : #FFFFFF;
-    --divColor: #FFFFFF;
-    --blueColor : #1A79FA;
-    --lableColor : #797979;
-  
-  }
-  
-  .darkTheme {
-    --bgColor : #424242;
-    --textColor : #FFFFFF;
-    --whiteColor : #FFFFFF;
-    --divColor: #555555;
-    --blueColor : #1A79FA;
-    --lableColor : #797979;
-  }
-
-
-  .historyGrid{
-    padding: 10px;
-    width: 100%;
-    background-color: var(--divColor);
-    border: 1px dashed var(--blueColor);
-    transition: 0.8s box-shadow;
-  }
-  .historyGrid:hover{
-    box-shadow:5px 5px 20px #1A79FA;
-  }
-.webhookTitle{
-  display: flex;
-  justify-content: space-between;
-}
-.copyResponse{
-  background-color: var(--blueColor);
-  color: var(--whiteColor);
-  border: none;
-  padding: 2px 15px;
-  border-radius: 10px;
-  cursor: pointer;
-
-}
-
-  .WebListBox{
-    /* background-color: #1A79FA; */
-    width: 200px;
-    height: 100%;
-    overflow: auto;
-    
-  }
-
-  .webdetails{
-    color: var(--textColor);
-    text-align: center;
-  }
-  .receivedTime{
-    display: flex;
-    align-items: center;
-    gap: 10px;
-
-  }
+﻿
